+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * optimize: `setHighlightDimRow` 使用 `document.getElementById` 实现。优化渲染性能。
 * optimize: StkTable.vue 中 变量`tableHeader`/`tableHeaderLast` 转为 `shallowRef` 优化
 * fix: **非**虚拟滚动时，键盘上下按键及`PageUp`/`PageDown`无法滚动问题。
-* fix: 修复 **非**横向虚拟滚动时，列宽不设置为变为默认列宽的问题。
+* fix: **非**横向虚拟滚动时，列宽不设置为变为默认列宽的问题。
+* fix: **非**虚拟滚动时，列固定计算应优先使用 `minWidth` 计算，非 `width`。
 
 ## 0.2.9
 * feature: 新增`props.highlightConfig` 支持配置高亮参数（持续时间，颜色，频率（虚拟滚动））。